+++ backend/db/util/insert_or_update_container.py
@@ -14,6 +14,9 @@ class ContainerInsertOrUpdateData(TypedDict, total=False):
     update_available: bool
     checked_at: datetime
     updated_at: datetime
+    local_digests: list[str]
+    remote_digests: list[str]
+    image_id: str
 
 
 async def insert_or_update_container(